Implement grab-fn-call to enable other debug tools
Grab call closes over a fn and grab its meta, args and result, being
just a return value or a exception, and return in a map. This fn will
enable us to build the other debug tools.

Besides that we also added deps.edn and bin/kaocha to run our test.