Capturing exceptions and marking captures to better reporting later
We want to have better sinks in the future, but for that we need to know
what came from grasp and what not, so we are adding a meta data to
values tapped so we can filter them later.

Capturing exceptions as well as values.

Adding simple pprint and rebl sinks.

Also renaming tap -> grab just because.