Implementing defn replacement
This replacement will grab the parameters and the return.
It takes care of some of the different ways you can define a function,
but not all of them, just the ones that I use the most, but if we need
for some that is not covered it can be improved later.